Cleanup of exchange.js
